Started Mocking Presenter
I have created all the mock variables for the presenter aswell as all of the mock methods i want to test. I have commented the functionality of each of these methods so they are well defined when i come to test. Dependency injection is going smoothly and i have a well defined interface for every part of the MVP design pattern.